Move AppCacheUpdateWorker.kt to appmanager/mobile
