+++ appmanager/mobile/src/androidMain/kotlin/com/boswelja/smartwatchextensions/appmanager/AppCacheUpdateWorker.kt
@@ -10,6 +10,9 @@ import androidx.work.WorkManager
 import androidx.work.WorkerParameters
 import androidx.work.workDataOf
 import com.boswelja.watchconnection.common.Watch
+import com.boswelja.watchconnection.common.message.Message
+import com.boswelja.watchconnection.common.message.MessageClient
+import com.boswelja.watchconnection.serialization.MessageHandler
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.map
 import org.koin.core.component.KoinComponent
@@ -20,42 +23,46 @@ import java.util.concurrent.TimeUnit
 /**
  * A [CoroutineWorker] designed to validate app cache for the given watch.
  */
-abstract class BaseAppCacheUpdateWorker(
+class AppCacheUpdateWorker(
     context: Context,
     workerParams: WorkerParameters
 ) : CoroutineWorker(context, workerParams), KoinComponent {
 
     private val appRepository: WatchAppRepository by inject()
-
-    /**
-     * Called when the cache hash is ready to be sent. Sending logic should be handled here.
-     * @param targetUid The target device UID.
-     * @param cacheHash The cache hash.
-     */
-    abstract suspend fun onSendCacheState(
-        targetUid: String,
-        cacheHash: Int
-    ): Boolean
+    private val messageClient: MessageClient by inject()
 
     override suspend fun doWork(): Result {
         val watchId = inputData.getString(EXTRA_WATCH_ID)
-        if (watchId.isNullOrBlank()) {
-            return Result.failure()
-        }
+        requireNotNull(watchId)
 
         // Get cache hash
-        val apps = appRepository.getAppVersionsFor(watchId)
-            .map { apps ->
-                apps.map { it.packageName to it.updateTime }
-            }
-            .first()
-        val cacheHash = CacheValidation.getHashCode(apps)
+        val cacheHash = calculateCacheHash(watchId)
 
-        return if (onSendCacheState(watchId, cacheHash)) {
+        return if (sendCacheState(watchId, cacheHash)) {
             Result.success()
         } else {
             Result.retry()
         }
+    }
+
+    private suspend fun calculateCacheHash(targetUid: String): Int {
+        val apps = appRepository.getAppVersionsFor(targetUid)
+            .map { apps ->
+                apps.map { it.packageName to it.updateTime }
+            }
+            .first()
+        return CacheValidation.getHashCode(apps)
+    }
+
+    private suspend fun sendCacheState(targetUid: String, cacheHash: Int): Boolean {
+        val handler = MessageHandler(CacheValidationSerializer, messageClient)
+        return handler.sendMessage(
+            targetUid,
+            Message(
+                VALIDATE_CACHE,
+                cacheHash
+            )
+        )
     }
 
     companion object {
@@ -66,23 +73,23 @@ abstract class BaseAppCacheUpdateWorker(
         const val EXTRA_WATCH_ID: String = "extra_watch_id"
 
         /**
-         * Get a string that can be used to represent an [BaseAppCacheUpdateWorker] for a watch with
-         * a given ID.
+         * Get a string that can be used to represent an [AppCacheUpdateWorker] for a watch with a
+         * given ID.
          * @param watchId The [Watch.uid] of the watch associated with the worker.
          * @return A string unique to the worker defined for the watch with the given ID.
          */
-        fun getWorkerNameFor(watchId: String): String {
+        private fun getWorkerNameFor(watchId: String): String {
             return "appcache-$watchId"
         }
 
         /**
-         * Enqueue a new [BaseAppCacheUpdateWorker] with appropriate constraints for the watch with
-         * the given ID.
+         * Enqueue a new [AppCacheUpdateWorker] with appropriate constraints for the watch with the
+         * given ID.
          * @param context [Context].
          * @param watchId The [Watch.uid] of the watch associated with the worker.
          * @return The ID of the work request.
          */
-        inline fun <reified T : BaseAppCacheUpdateWorker> enqueueWorkerFor(
+        fun enqueueWorkerFor(
             context: Context,
             watchId: String
         ): UUID {
@@ -96,7 +103,7 @@ abstract class BaseAppCacheUpdateWorker(
             val data = workDataOf(EXTRA_WATCH_ID to watchId)
 
             // Create a work request
-            val request = PeriodicWorkRequestBuilder<T>(
+            val request = PeriodicWorkRequestBuilder<AppCacheUpdateWorker>(
                 1, TimeUnit.DAYS
             ).setConstraints(constraints)
                 .setInputData(data)
@@ -114,7 +121,7 @@ abstract class BaseAppCacheUpdateWorker(
         }
 
         /**
-         * Stops any enqueued [BaseAppCacheUpdateWorker] for the watch with the given ID.
+         * Stops any enqueued [AppCacheUpdateWorker] for the watch with the given ID.
          * @param context [Context].
          * @param watchId The [Watch.uid] of the watch associated with the worker.
          */
